requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
PyPDF2>=3.0.1
python-docx>=1.1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
google-generativeai>=0.3.2
httpx>=0.25.2
python-dotenv>=1.0.0
aiofiles>=23.2.1
langchain>=0.1.0
langchain-community>=0.0.10
numpy>=1.24.3
